move latest person
